Move read_participant_tsv, make it private and change associated tests
