+++ test/unittests/utils/test_filemanip.py
@@ -218,43 +218,6 @@ def test_extract_crash_files_from_log_file_error():
         extract_crash_files_from_log_file("foo.log")
 
 
-def test_read_participant_tsv_error(tmp_path):
-    from clinica.utils.exceptions import ClinicaException
-    from clinica.utils.filemanip import read_participant_tsv
-
-    with pytest.raises(
-        ClinicaException,
-        match="The TSV file you gave is not a file.",
-    ):
-        read_participant_tsv(tmp_path / "foo.tsv")
-
-
-def test_read_participant_tsv(tmp_path):
-    from clinica.utils.exceptions import ClinicaException
-    from clinica.utils.filemanip import read_participant_tsv
-
-    df = pd.DataFrame(
-        {
-            "participant_id": ["sub-01", "sub-01", "sub-02"],
-            "session_id": ["ses-M000", "ses-M006", "ses-M000"],
-        }
-    )
-    df.to_csv(tmp_path / "foo.tsv", sep="\t")
-
-    assert read_participant_tsv(tmp_path / "foo.tsv") == (
-        ["sub-01", "sub-01", "sub-02"],
-        ["ses-M000", "ses-M006", "ses-M000"],
-    )
-
-    for column in ("participant_id", "session_id"):
-        df.drop(column, axis=1).to_csv(tmp_path / "foo.tsv", sep="\t")
-        with pytest.raises(
-            ClinicaException,
-            match=f"The TSV file does not contain {column} column",
-        ):
-            read_participant_tsv(tmp_path / "foo.tsv")
-
-
 def test_extract_metadata_from_json_missing_file_error(tmp_path):
     from clinica.utils.filemanip import extract_metadata_from_json
